Put copied registers in a separate file
