+++ CMakeLists.txt
@@ -55,6 +55,7 @@ set(SOURCES
         src/core/gif.cpp
         src/core/gs.cpp
         src/core/gsthread.cpp
+        src/core/gsregisters.cpp
         src/core/gscontext.cpp
 	src/core/sif.cpp
 	src/qt/emuthread.cpp
@@ -103,6 +104,7 @@ set(HEADERS
         src/core/gif.hpp
         src/core/gs.hpp
         src/core/gsthread.hpp
+        src/core/gsregisters.hpp
         src/core/circularFIFO.hpp
 	src/core/gscontext.hpp
 	src/core/int128.hpp
