+++ DobieStation/DobieStation.pro
@@ -23,6 +23,7 @@ SOURCES += ../src/qt/main.cpp \
     ../src/core/ee/bios_hle.cpp \
     ../src/core/ee/emotion_special.cpp \
     ../src/core/gs.cpp \
+    ../src/core/gsregisters.cpp \
     ../src/core/gsthread.cpp \
     ../src/core/ee/dmac.cpp \
     ../src/qt/emuwindow.cpp \
@@ -74,6 +75,7 @@ HEADERS += \
     ../src/core/gs.hpp \
     ../src/core/circularFifo.hpp \
     ../src/core/gsthread.hpp \
+    ../src/core/gsregisters.hpp \
     ../src/core/ee/dmac.hpp \
     ../src/qt/emuwindow.hpp \
     ../src/core/gscontext.hpp \
